refatoração para que o service de consolidação se adeque somente ao contexto de atualização de notificação
